requirements.dev.txt
# Non-production dependencies
pydantic-extra-types
black
isort
